Add ApiJobKind, rename module
