+++ qiskit/providers/ibmq/ibmqjob.py
@@ -32,7 +32,7 @@ from qiskit.qobj import Qobj, validate_qobj_against_schema
 from qiskit.result import Result
 
 from .api import ApiError
-from .apijobstatus import ApiJobStatus
+from .apiconstants import ApiJobStatus
 from .api_v2.exceptions import WebsocketTimeoutError, WebsocketError
 
 logger = logging.getLogger(__name__)
