+++ test/ibmq/test_ibmq_job_states.py
@@ -19,7 +19,7 @@
 import time
 from contextlib import suppress
 
-from qiskit.providers.ibmq.apijobstatus import API_JOB_FINAL_STATES, ApiJobStatus
+from qiskit.providers.ibmq.apiconstants import API_JOB_FINAL_STATES, ApiJobStatus
 from qiskit.test.mock import new_fake_qobj, FakeRueschlikon
 from qiskit.providers import JobError, JobTimeoutError
 from qiskit.providers.ibmq.api import ApiError
